added campaigns

+++ fundsphere-server/controllers/admin.controller.js
@@ -1,4 +1,5 @@
 const User = require("../models/user.model");
+const Campaign = require("../models/campaign.model");
 
 const getAllUsers = async (req, res, next) => {
   try {
@@ -84,15 +85,25 @@ const deleteUsersByAdmin = async (req, res, next) => {
       return res.status(400).json({ msg: "No users selected" });
     }
 
-    const result = await User.deleteMany({
+    const users = await User.Find({
       _id: { $in: ids },
       role: { $ne: "admin" },
     });
 
+    if (!users.length)
+      return res.status(404).json({ error: "No deletable users found" });
+
+    const userIds = users.map((u) => u._id);
+
+    // Delete linked campaigns
+    await Campaign.deleteMany({ createdBy: { $in: userIds } });
+
+    const result = await User.deleteMany({ _id: { $in: userIds } });
+
     res.json({
       msg: `Deleted ${result.deletedCount} ${
         result.deletedCount === 1 ? "user" : "users"
-      } successfully.`,
+      } and their campaigns successfully.`,
     });
   } catch (error) {
     next(error);
@@ -110,10 +121,12 @@ const deleteUserById = async (req, res, next) => {
     if (user.role === "admin")
       return res.status(403).json({ msg: "Cannot delete an admin" });
 
+    await Campaign.deleteMany({ createdBy: user._id });
+
     await user.deleteOne();
 
     res.json({
-      msg: `User ${user.username} deleted successfully.`,
+      msg: `User ${user.username} and their campaigns deleted successfully.`,
     });
   } catch (error) {
     next(error);
